made changes to testing directory
